[Mod] 修正ci脚本
