+++ tools/format_all.py
@@ -5,11 +5,11 @@ import subprocess
 logger = logging.Logger(__file__)
 
 
-def format_by_black():
-    subprocess.check_call(["black", "vnpy"])
+def format_black():
+    subprocess.check_call(["python", "-m", "black", "."])
 
 
 if __name__ == "__main__":
-    if not format_by_black():
+    if not format_black():
         exit(1)
     exit(0)
